Create README.md
This is my first personal project that I build using HTML, CSS and JavaScript. It is a tip calculator that you can plug in your bill amount, use the slider to the desired tip percentage and it will tell you how much the tip will be and how much the total will come out to.